move more logic and helper methods into AccountManager
